requirements.txt
streamlit
groq
google-generativeai
requests
beautifulsoup4
python-dotenv
pandas